requirements.txt
absl-py==2.4.0
aiofiles==24.1.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
axial_positional_embedding==0.3.12
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clip @ git+https://github.com/ultralytics/CLIP.git@88ade288431a46233f1556d1e141901b3ef0a36b
colorama==0.4.6
coloredlogs==15.0.1
CoLT5-attention==0.11.1
ConfigArgParse==1.7.5
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.4
decord==0.6.0
deep-sort-realtime==1.3.2
easydict==1.13
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.135.1
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.1.0
ftfy==6.3.1
gradio==6.11.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.2
gdown==6.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
humanfriendly==10.0
hyper-connections==0.4.9
idna==3.11
ImageIO==2.37.3
insightfacex==0.7.4
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy-loader==0.5
linear-attention-transformer==0.19.1
linformer==0.2.3
local-attention==1.11.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
onnx==1.20.1
onnxruntime-gpu==1.23.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
orjson==3.11.8
packaging==26.0
pandas==2.3.3
pillow==12.1.0
pillow-heif==1.3.0
polars==1.37.1
polars-runtime-32==1.37.1
prettytable==3.17.0
product_key_memory==0.3.0
protobuf==4.25.8
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python_speech_features==0.6
Pygments==2.19.2
pyparsing==3.3.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.3
safehttpx==0.1.7
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
scenedetect==0.6.7.1
semantic-version==2.10.0
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
sounddevice==0.5.5
starlette==0.52.1
stringzilla==4.6.0
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.5.1+cu121
torch-einops-utils==0.0.30
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
yt-dlp==2026.2.4
